% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmbart.R
\name{hmbart}
\alias{hmbart}
\title{Hierarchical Mediation Analysis using BART}
\usage{
hmbart(
  data,
  X,
  t,
  m,
  y,
  CV = FALSE,
  num_trees = 100,
  num_tree_cvs = c(100),
  k_cvs = c(2, 3, 5),
  nu_q_cvs = list(c(3, 0.9), c(3, 0.99), c(10, 0.75)),
  n_burn_in = 2000,
  n_after_burn_in = 500,
  n_process_samples = 1e+05
)
}
\arguments{
\item{data}{A data frame containing the dataset to be analyzed.}

\item{X}{A character vector specifying the names of covariates to include in the model.}

\item{t}{A character string specifying the name of the treatment variable.}

\item{m}{A character string specifying the name of the mediator variable.}

\item{y}{A character string specifying the name of the outcome variable.}

\item{CV}{Logical; whether to use cross-validation to tune BART hyperparameters. Default is FALSE.}

\item{num_trees}{Number of trees for BART models. Default is 100.}

\item{num_tree_cvs}{A vector specifying the number of trees to consider during cross-validation. Default is c(100).}

\item{k_cvs}{A vector specifying the k parameter values for BART cross-validation. Default is c(2, 3, 5).}

\item{nu_q_cvs}{A list specifying the prior degrees of freedom and quantiles for the BART error distribution in cross-validation. Default is list(c(3, 0.9), c(3, 0.99), c(10, 0.75)).}

\item{n_burn_in}{Number of burn-in iterations for BART. Default is 2000.}

\item{n_after_burn_in}{Number of post burn-in iterations for BART. Default is 500.}

\item{n_process_samples}{Number of samples to process during posterior sampling. Default is 1e5.}
}
\value{
A list containing:
\item{data}{The processed dataset with standardized mediator and outcome variables.}
\item{X_name}{The names of covariates included in the model.}
\item{effects}{A data frame summarizing treatment effects (TE), natural direct effects (NDE), and natural indirect effects (NIE) along with their 95\% credible intervals.}
}
\description{
This function implements a hierarchical mediation analysis framework using Bayesian Additive Regression Trees (BART).
It estimates treatment effects (TE), natural direct effects (NDE), and natural indirect effects (NIE) along with credible intervals.
}
\details{
The function uses the \code{bartMachine} package to fit BART models for the mediator and outcome variables.
It supports cross-validation to tune hyperparameters and computes posterior samples to derive causal effects.
}
